Add routes + login test.
